BACKEND/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
python-dotenv==1.0.0
webrtcvad==2.0.10
faster-whisper==1.0.3
groq==0.11.0
requests==2.31.0
pymupdf==1.23.10
sentence-transformers==2.7.0
numpy>=1.26.0
pydub==0.25.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
tavily-python==0.3.3
elevenlabs==0.2.27